Do some refactoring to make code more updateable
Move setting query to abstract method and made ExecuteQuery method returning DataTable type. If there is nothing to return, for example in insert query we can return null. It is not ideal but for now - ok. We have one base class and for every query we can simply override SetQuery function

Add fundamentals for delete rental procedure