Finish code for deleting rentals functionality
Change a little way of programming procedure classes. Now I am returniing DataTable with affected rows number. And again, not great idea but for now okay.  I am checking if returned table is empty or not and I can tell to user that procedure changed something or no. If no, there is something wrong with provided data